Refactored and cleaned code
